Move platform independent data classes
