+++ api-internal/src/main/java/at/pavlov/internal/cannon/data/CannonMainData.java
@@ -1,4 +1,4 @@
-package at.pavlov.cannons.cannon.data;
+package at.pavlov.internal.cannon.data;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
