+++ cannons-bukkit/src/main/java/at/pavlov/cannons/cannon/Cannon.java
@@ -5,14 +5,14 @@ import at.pavlov.cannons.Enum.BreakCause;
 import at.pavlov.cannons.Enum.CannonRotation;
 import at.pavlov.cannons.Enum.InteractAction;
 import at.pavlov.cannons.Enum.MessageEnum;
-import at.pavlov.cannons.cannon.data.AimingData;
+import at.pavlov.internal.cannon.data.AimingData;
 import at.pavlov.cannons.cannon.data.AmmoLoadingData;
-import at.pavlov.cannons.cannon.data.AngleData;
-import at.pavlov.cannons.cannon.data.CannonMainData;
+import at.pavlov.internal.cannon.data.AngleData;
+import at.pavlov.internal.cannon.data.CannonMainData;
 import at.pavlov.cannons.cannon.data.CannonPosition;
-import at.pavlov.cannons.cannon.data.LinkingData;
-import at.pavlov.cannons.cannon.data.SentryData;
-import at.pavlov.cannons.cannon.data.WhitelistData;
+import at.pavlov.internal.cannon.data.LinkingData;
+import at.pavlov.internal.cannon.data.SentryData;
+import at.pavlov.internal.cannon.data.WhitelistData;
 import at.pavlov.cannons.container.ItemHolder;
 import at.pavlov.cannons.container.SimpleBlock;
 import at.pavlov.cannons.cannon.data.FiringData;
