+++ cannons-bukkit/src/main/java/at/pavlov/cannons/interfaces/holders/CannonMainDataHolder.java
@@ -1,6 +1,6 @@
 package at.pavlov.cannons.interfaces.holders;
 
-import at.pavlov.cannons.cannon.data.CannonMainData;
+import at.pavlov.internal.cannon.data.CannonMainData;
 import at.pavlov.cannons.interfaces.functionalities.Updatable;
 
 import java.util.UUID;
