+++ cannons-bukkit/src/main/java/at/pavlov/cannons/interfaces/holders/LinkingDataHolder.java
@@ -1,8 +1,7 @@
 package at.pavlov.cannons.interfaces.holders;
 
 import at.pavlov.cannons.Enum.MessageEnum;
-import at.pavlov.cannons.cannon.data.LinkingData;
-import org.apache.commons.lang3.Validate;
+import at.pavlov.internal.cannon.data.LinkingData;
 import org.bukkit.entity.Player;
 
 import java.util.UUID;
